feat: verify and login with Server Action
