+++ .changeset/breezy-seas-heal.md
@@ -1,0 +1,5 @@
+---
+"@mogeko/bookworm-web": patch
+---
+
+Verify and Login with [Server Actions](https://nextjs.org/docs/app/building-your-application/data-fetching/server-actions) instead of [REST API](https://github.com/mogeko/bookworm/blob/eb675825116860414e6a2e2b9e8bb4716cb9a5ff/apps/web/app/auth/route.ts#L7).
